requirements.txt
pytest==7.2.1
selenium==4.8.0
webdriver-manager==3.8.5
allure-pytest==2.12.0
allure-python-commons==2.12.0
requests==2.28.2
